refactor: Add message deletion on moderation actions (ban, mute, warn)
